# [SPARK-40465][SQL] Refactor Decimal so as we can use other underlying implementation