--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/types/DecimalOperation.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/types/DecimalOperation.scala
@@ -209,13 +209,13 @@ trait DecimalOperation[T <: DecimalOperation[T]] extends Serializable {
   def toScalaBigInt: BigInt = if (isNotNull) {
     getAsBigDecimal().toBigInt
   } else {
-    BigInt(toLong)
+    BigInt(actualLongVal)
   }
 
   def toJavaBigInteger: java.math.BigInteger = if (isNotNull) {
     getAsJavaBigInteger()
   } else {
-    java.math.BigInteger.valueOf(toLong)
+    java.math.BigInteger.valueOf(actualLongVal)
   }
 
   protected def getAsJavaBigInteger(): java.math.BigInteger
@@ -233,10 +233,12 @@ trait DecimalOperation[T <: DecimalOperation[T]] extends Serializable {
   }
 
   def toLong: Long = if (isNull) {
-    longVal / POW_10(_scale)
+    actualLongVal
   } else {
     longValue
   }
+
+  private def actualLongVal: Long = longVal / POW_10(_scale)
 
   protected def longValue: Long
 
@@ -246,7 +248,6 @@ trait DecimalOperation[T <: DecimalOperation[T]] extends Serializable {
       maxValue: Int,
       minValue: Int) (f1: Long => T) (f2: Double => T): T = {
     if (isNull) {
-      val actualLongVal = longVal / POW_10(_scale)
       val numericVal = f1(actualLongVal)
       if (actualLongVal == numericVal) {
         numericVal
@@ -266,7 +267,7 @@ trait DecimalOperation[T <: DecimalOperation[T]] extends Serializable {
   }
 
   def roundToLong(decimal: Decimal): Long = if (isNull) {
-    longVal / POW_10(_scale)
+    actualLongVal
   } else {
     try {
       // We cannot store Long.MAX_VALUE as a Double without losing precision.
@@ -344,7 +345,7 @@ trait DecimalOperation[T <: DecimalOperation[T]] extends Serializable {
     if (isNotNull) {
       // We get here if either we started with a BigDecimal, or we switched to one because we would
       // have overflowed our Long; in either case we must rescale dv to the new scale.
-      if (!rescale(scale, roundMode)) {
+      if (!rescale(precision, scale, roundMode)) {
         return false
       }
     } else {
@@ -361,7 +362,8 @@ trait DecimalOperation[T <: DecimalOperation[T]] extends Serializable {
     true
   }
 
-  protected def rescale(scale: Int, roundMode: BigDecimal.RoundingMode.Value): Boolean
+  protected def rescale(
+      precision: Int, scale: Int, roundMode: BigDecimal.RoundingMode.Value): Boolean
 
   def compare(that: DecimalOperation[_]): Int =
     if (isNull && that.isNull && _scale == that._scale) {