@beliefer 2022-09-20
This PR can ensure that the existing underlying implementation of `Decimal` will not be changed, and it also provides an interface for us to develop new implementations through iteration. With this interface, no matter whether we can enter the community in the future, our development and maintenance costs will be greatly reduced.