@beliefer 2022-09-21
> The abstraction makes sense to me. The new `DecimalOperation` layer adds an indirection and may hurt performance, though I think the overhead is minor as Java decimal operation is already slow. Can we add a microbenchmark for `Decimal`? You can follow `HashBenchmark`.

I followed `HashBenchmark` and create a new benchmark `DecimalBenchmark`.